requirements.txt
fastapi
uvicorn
openfga-sdk
httpx
python-dotenv